==> 4. Cohort Comparison - Logistic Regression - Finnish.Rmd <==
---
title: "Low ferritin Logistic model"
author: "Sofie Ekroos"
date: "7.11.2021"
output:
  html_document: 
    df_print: paged
---
 
# Summary

Code and documentation for running logistic regression of iron deficiency (ferritin <15 mcg/l) for the Finnish cohorts (FinDonor, FINRISK97, Health 2000) in the paper "Iron status in Dutch and Finnish blood donors and general populations: a cross-cohort comparison study". 

This code includes both non-scaled and scaled Bayesian logistic regression models. 

```{r setup, include=FALSE}
Sys.setlocale("LC_MESSAGES", 'en_GB.UTF-8')
Sys.setenv(LANG = "en_US.UTF-8")

knitr::opts_chunk$set(
        echo = FALSE,
        message = FALSE,
        warning = FALSE)

currentDate <- format(Sys.time(), "%Y-%m-%d")
FolderName <- paste("../results/genpop_combined/low_ferritin_model_bayes", currentDate, sep="")
dir.create(FolderName)


library(lubridate)
library(plyr)
library(readr)
#library(nFactors)
#library(psych)
#library(corrplot)
#library(psych)
library(ggplot2)
library(ggthemes)
library(haven)
library(rsample)
library(broom)
library(scales)
library(kableExtra)
library(naniar)
library(haven)
library(foreign)
library(tableone)
library(stargazer)
library(gridExtra)
library(cowplot)
library(broom)
library(GGally)
library(ggfortify)
library(knitr)
library(lubridate)
library(car)
library(reshape2)
library(sfsmisc)
library(MASS)
library(ordinal)
library(sjmisc)
#library(VIM)
library(epiDisplay)
library(epitools)
library(ez)
library(ggbeeswarm)
library(ggthemes)
library(rsample)
library(broom)
library(scales)
library(kableExtra)
library(boot)
library(readxl)
library(foreign)
#library(corrr)
#library(ggcorrplot)
library(tidyverse)
library(brms)
library(bayesplot)
library(tidylog, warn.conflicts = F)
```

BAYESIAN LOGISTIC REGRESSION

# Data loading and preparation 
```{r}
load("~/proj/iron_measurement_comparisons/data/ID_data_regression_cohorts_genpop_combined.rdata")
all_cohorts <- regression_cohorts %>% 
  mutate_if(is.ordered, function(x){factor(x, ordered = FALSE)}) # remove ordering

# Rename THL 
all_cohorts$Cohort <- gsub("THL", "General Population", all_cohorts$Cohort)

# Rename NEW
all_cohorts$Cohort <- gsub("FINDONOR", "Blood donors", all_cohorts$Cohort)

```

# Rename variables en create log scale variables to use in this script
```{r}
names(all_cohorts)[names(all_cohorts) == "TwoYearsFromStartCount_FB"] <- "donation_count"
names(all_cohorts)[names(all_cohorts) == "DaysToPreviousFB"] <- "last_donation"
names(all_cohorts)[names(all_cohorts) == "Pregnancy"] <- "PreviousChildbirth"

all_cohorts$Group <- as.character(all_cohorts$Group)


all_cohorts$log_CRP = log(all_cohorts$CRP)
all_cohorts$log_ferritin = log(all_cohorts$Ferritin)
all_cohorts$log_last_donation = log(all_cohorts$last_donation)


# all_cohorts$Smoking <- gsub("Nee, maar ik heb in het verleden wel gerookt", "No",all_cohorts$Smoking)
# all_cohorts$Smoking <- gsub("Nee, nooit gerookt", "No",all_cohorts$Smoking)
# all_cohorts$Smoking <- gsub("Nee, nooit", "No",all_cohorts$Smoking)
# all_cohorts$Smoking <- gsub("Nee, maar wel gerookt", "No",all_cohorts$Smoking)
# all_cohorts$Smoking <- gsub("Ja", "Yes",all_cohorts$Smoking)
# 
# all_cohorts$Menstruation <- gsub("Ja", "Yes",all_cohorts$Menstruation)
# all_cohorts$Menstruation <- gsub("Nee", "No",all_cohorts$Menstruation)
# all_cohorts$Menstruation <- gsub("No, not for longer than 10 years", "No",all_cohorts$Menstruation)
# all_cohorts$Menstruation <- gsub("No, not since 1-2 years", "No",all_cohorts$Menstruation)
# all_cohorts$Menstruation <- gsub("No, not since 3-5 years", "No",all_cohorts$Menstruation)
# all_cohorts$Menstruation <- gsub("No, not since 6-10 years", "No",all_cohorts$Menstruation)
# all_cohorts$Menstruation <- gsub("No, not since a few months", "No",all_cohorts$Menstruation)
# all_cohorts$Menstruation <- gsub("No, not since a few months", "No",all_cohorts$Menstruation)
# all_cohorts$Menstruation <- gsub("1", "Yes",all_cohorts$Menstruation)
# all_cohorts$Menstruation <- gsub("2", "No",all_cohorts$Menstruation)
# all_cohorts$Menstruation <- gsub("3", "No",all_cohorts$Menstruation)
# all_cohorts$Menstruation <- gsub("4", "No",all_cohorts$Menstruation)
# all_cohorts$PreviousChildbirth <- gsub("Ja", "Yes",all_cohorts$PreviousChildbirth)
# all_cohorts$PreviousChildbirth <- gsub("Nee", "No",all_cohorts$PreviousChildbirth)
# all_cohorts$Smoking <- gsub("Missing", NA,all_cohorts$Smoking)


```

## Data
```{r}
# men
## menstruation and childbirth variables removed 
regression_data_men <- all_cohorts %>% 
  filter(Group == "Men") %>% 
  dplyr::select(ID, log_ferritin, Age, log_CRP, blood_vol, donation_count, log_last_donation, Cohort, iron_deficiency, Smoking, Group, Weight) %>%
  mutate(Age = scale(Age, scale = FALSE)[,1],
         log_CRP = scale(log_CRP, scale = FALSE)[,1],
         donation_count = scale(donation_count, scale = FALSE)[,1],
         log_last_donation = scale(log_last_donation, scale = FALSE)[,1],
         blood_vol = scale(blood_vol, scale = FALSE)[,1],
         donation_count_2 = donation_count^2) 

# premenopausal women
regression_data_pre <- all_cohorts %>% 
  filter(Group == "Pre_menopause_women") %>% 
  dplyr::select(ID, log_ferritin, Age, log_CRP, blood_vol, donation_count, log_last_donation, Cohort, iron_deficiency, Smoking, Group, Weight, HormonalContraception, PreviousChildbirth, Menstruation) %>%
  mutate(Age = scale(Age, scale = FALSE)[,1],
         log_CRP = scale(log_CRP, scale = FALSE)[,1],
         donation_count = scale(donation_count, scale = FALSE)[,1],
         log_last_donation = scale(log_last_donation, scale = FALSE)[,1],
         blood_vol = scale(blood_vol, scale = FALSE)[,1],
         donation_count_2 = donation_count^2) 

# postmenopausal womem 
## menstruation variable removed 
regression_data_post <- all_cohorts %>% 
  filter(Group == "Post_menopause_women") %>% 
  dplyr::select(ID, log_ferritin, Age, log_CRP, blood_vol, donation_count, log_last_donation, Cohort, iron_deficiency, Smoking, Group, Weight, HormonalContraception, PreviousChildbirth, Menstruation) %>%
  mutate(Age = scale(Age, scale = FALSE)[,1],
         log_CRP = scale(log_CRP, scale = FALSE)[,1],
         donation_count = scale(donation_count, scale = FALSE)[,1],
         log_last_donation = scale(log_last_donation, scale = FALSE)[,1],
         blood_vol = scale(blood_vol, scale = FALSE)[,1],
         donation_count_2 = donation_count^2) 


```

# Model 1 [Crude Model]


# Model 1: Men
```{r men m1}
#men
logistic_men_b <- brm(iron_deficiency ~ 
                        #Age + 
                        #blood_volume + 
                        #Smoking + 
                        # donation_count_2 + 
                        # donation_count +
                        # log_last_donation + 
                        #log_CRP +
                        #Weight +
                        Cohort, 
                      data = regression_data_men, 
                      family = bernoulli(),
                      #file=file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_men_b)



```


# Model 1: Premenopausal Women
```{r pre m1}

logistic_pre_b <- brm(iron_deficiency ~ 
                        #Age + 
                        #blood_volume + 
                        #Smoking + 
                        # donation_count_2 + 
                        # donation_count +
                        # log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort, #+
                        #Menstruation +
                        #PreviousChildbirth, 
                      data = regression_data_pre,
                      family = bernoulli(),
                      #file=file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_pre_b)



```


# Model 1: Postmenopausal Women
```{r post m1}

logistic_post_b <- brm(iron_deficiency ~
                        #Age + 
                        #blood_volume + 
                        # donation_count_2 +
                        # donation_count +
                        # log_last_donation + 
                        #Weight +
                        Cohort,
                        #log_CRP +
                        #Smoking + 
                        #RedMeat_n +
                        #+
                        #Menstruation +   
                        #PreviousChildbirth, 
                     data = regression_data_post, 
                     family = bernoulli(),
                     #file=file, 
                     cores = 4,
                     iter= 10000
)

summary(logistic_post_b)



```








## Diagnostics

### Trace

```{r , message=FALSE}
mcmc_plot(logistic_men_b, type = "trace")
```

```{r , message=FALSE}
mcmc_plot(logistic_pre_b, type = "trace")
```

```{r , message=FALSE}
mcmc_plot(logistic_post_b, type = "trace")
```

Chains do not seem to get stuck.

### ACF bars

```{r}
mcmc_plot(logistic_men_b, type = "acf_bar")
```


```{r}
mcmc_plot(logistic_pre_b, type = "acf_bar")
```


```{r}
mcmc_plot(logistic_post_b, type = "acf_bar")
```

Autocorrelation drops nicely.

### Rhat

```{r}
mcmc_plot(logistic_men_b, type = "rhat")
```

```{r}
mcmc_plot(logistic_pre_b, type = "rhat")
```

```{r}
mcmc_plot(logistic_post_b, type = "rhat")
```
https://mc-stan.org/rstan/reference/Rhat.html


"The Rhat function produces R-hat convergence diagnostic, which compares the between- and within-chain estimates for model parameters and other univariate quantities of interest. If chains have not mixed well (ie, the between- and within-chain estimates don't agree), R-hat is larger than 1. We recommend running at least four chains by default and only using the sample if R-hat is less than 1.05."

# Distributions


```{r , message=FALSE, warning=FALSE}
p <- mcmc_plot(logistic_men_b, 
         type = "areas",
         prob = 0.95,
         transformations = "exp") +
  geom_vline(xintercept = 1, color = "grey")

p <- p + scale_x_log10()
#Ferritin needs to be removed
# plot.OR<-  tidy_logit_pre_menop_women %>%filter(!term=="log_Ferritin_beginning") %>%  dplyr::select(OR) %>% mutate(y=rev(2:(nrow(tidy_logit_pre_menop_women)))) %>% bind_rows(tibble(OR=NA,y=1 ))
# p <- p + geom_point(aes(y=y,x=OR),data=plot.OR, col="red")
p
```


```{r , message=FALSE, warning=FALSE}
p <- mcmc_plot(logistic_pre_b, 
         type = "areas",
         prob=0.95,
         prob_est="median",
         transformations = "exp"
        ) +
  geom_vline(xintercept = 1, color = "grey")

p <- p + scale_x_log10()
p
```


```{r , message=FALSE, warning=FALSE}
p <- mcmc_plot(logistic_post_b, 
         type = "areas",
         prob = 0.95,
         transformations = "exp") +
  geom_vline(xintercept = 1, color = "grey")

p <- p + scale_x_log10()
#Ferritin needs to be removed
# plot.OR<-  tidy_logit_pre_menop_women %>%filter(!term=="log_Ferritin_beginning") %>%  dplyr::select(OR) %>% mutate(y=rev(2:(nrow(tidy_logit_pre_menop_women)))) %>% bind_rows(tibble(OR=NA,y=1 ))
# p <- p + geom_point(aes(y=y,x=OR),data=plot.OR, col="red")
p
```

### All groups together

```{r}
regressor_values <- c(
    "Age" = "Age (5 years)",
    "CohortGeneralPopulation" = "General population", 
   "log_CRP" = "CRP (log transformed)",
   "blood_vol" = "Blood volume (Nadler)",
   "log_last_donation" = "Days since last donation (log, /log(2))",
   "Smokingyes" ="Smoking",
    "iron_complience" = "Iron supplementation",
    "donation_count_2" = "Donations in last 2 years (quadratic)",
    "Menstruationregular_period" = "Regular periods",
    "Weight" = "Weight (10 kg)",
    "PreviousChildbirth" = "Given birth in the past",
    "donation_count" = "Donations in last 2 years"
)
```



# Model 1: Intervals
```{r }

intervals_men <- mcmc_intervals_data(
  logistic_men_b,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Men')

 intervals_men <- intervals_men%>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  filter(! (parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

intervals_pre <- mcmc_intervals_data(
  logistic_pre_b,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Pre_menopause_women') %>% 
  arrange(m) # how to order the variables?

 intervals_pre <- intervals_pre%>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  filter(! (parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

intervals_post <- mcmc_intervals_data(
  logistic_post_b,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Post_menopause_women')

 intervals_post <- intervals_post%>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  filter(! (parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

intervals_combined <- rbind(intervals_pre,intervals_post,intervals_men) %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

```

# Model 1: Forest plot - Men, Premonpausal women, Postmenopausal women
```{r , fig1, fig.height = 10, fig.width = 15 ,warning=FALSE}

orlower <- -0.5
orupper <- 2.5

# Truncate ORs for plotting
plotdata <- intervals_combined %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )

#theme_set(bayesplot::theme_default())
pos <- position_nudge(y=as.numeric(as.factor(plotdata$Group))/5 -0.1) 

p <- ggplot(plotdata, aes(x = m, y = regressor, color = Group)) + 
  geom_point(position = position_dodge(width = -0.4),
             size=4) +
  geom_linerange(aes(xmin = ll, xmax = hh), 
                 position = position_dodge(width = -0.4),
                 size = 1, 
                 linetype = 1
                 ) +
  scale_color_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                     limits = c("Men",  "Post_menopause_women","Pre_menopause_women" )) +
  scale_fill_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                    limits = c("Men",  "Post_menopause_women","Pre_menopause_women" )) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower,orupper)  +   
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14)
  ) +
  guides(shape = "none") +
  theme_classic() +
  ylab("Coefficient") 

p

FileName <- paste(FolderName, "/Forest Plot - Logistic - Model 1.png",sep="")
ggsave(FileName, width = 10, height = 7)

```



# Model 2 [Corrected]

## Model 2 [Corrected]: Men

```{r men m2}
#men
logistic_men_corrected <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        #Smoking + 
                        #iron_complience + 
                        #donation_count_2 + 
                        #donation_count +
                        #log_last_donation + 
                        #log_CRP +
                        #Region + 
                        # Weight +
                        #RedMeat_n +
                        Cohort, 
                      data = regression_data_men, 
                      family = bernoulli(),
                      #file = file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_men_corrected)

```

## Model 2 [Corrected]: Premenopausal Women

```{r  pre m2}
# premenopausal women
#file <- "../results/bmodels/version3/linear_pre_b"

logistic_pre_corrected <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        #Smoking + 
                        #iron_complience + 
                        #donation_count_2 + 
                        #donation_count +
                        #log_last_donation + 
                        #log_CRP +
                        #Region + 
                        # Weight +
                        #RedMeat_n +
                        Cohort, # +
                        #Menstruation +
                        #PreviousChildbirth, 
                      data = regression_data_pre,
                      family = bernoulli(),
                      #file=file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_pre_corrected)


```

## Model 2 [Corrected]: Postmenopausal Women

```{r  post m2}
# postmenopausal women
#file <- "../results/bmodels/version3/linear_post_b"

logistic_post_corrected <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        #Smoking + 
                        #iron_complience + 
                        #donation_count_2 + 
                        #donation_count +
                        #log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort, #+
                        #PreviousChildbirth,
                     data = regression_data_post, 
                     family = bernoulli(),
                     #file=file, 
                     cores = 4,
                     iter= 10000
)

summary(logistic_post_corrected)
```

#  Model 2: Retrieve CI data
```{r }
# Group: Men
intervals_men_corrected <- mcmc_intervals_data(
  logistic_men_corrected,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
) %>% mutate(Group='Men')

 intervals_men_corrected <- intervals_men_corrected%>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  filter(! (parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

# Group: Premenopausal women
intervals_pre_corrected <- mcmc_intervals_data(
  logistic_pre_corrected,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  )  %>% mutate(Group='Pre_menopausal_women')

 intervals_pre_corrected <- intervals_pre_corrected%>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  filter(! (parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

# Group: Postmenopausal women
intervals_post_corrected <- mcmc_intervals_data(
  logistic_post_corrected,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Post_menopausal_women')

 intervals_post_corrected <- intervals_post_corrected%>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  filter(! (parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

# Group: Combined
intervals_combined_corrected <- rbind(intervals_pre_corrected,intervals_post_corrected,intervals_men_corrected) %>% 
  filter(! (parameter == 'b_Intercept' | parameter == 'lp__'  | parameter == 'lprior' | parameter == 'sigma')) %>% 
  mutate(parameter = gsub("b_","",parameter)) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values))

intervals_combined_corrected <- rbind(intervals_pre_corrected,intervals_post_corrected,intervals_men_corrected) %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  filter(! (parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

```

# Model 2: Forest plot  [Men, Premonpausal women, Postmenopausal women]
```{r , fig2, fig.height = 10, fig.width = 15 ,warning=FALSE}

orlower <- -0.5
orupper <- 2.5

# Truncate ORs for plotting
plotdata_combined_corrected <- intervals_combined_corrected %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )

#theme_set(bayesplot::theme_default())
pos <- position_nudge(y=as.numeric(as.factor(plotdata$Group))/5 -0.1) 

p <- ggplot(plotdata_combined_corrected, aes(x = m, y = regressor, color = Group)) + 
  geom_point(position = position_dodge(width = -0.5),
             size=4) +
  geom_linerange(aes(xmin = ll, xmax = hh), 
                 position = position_dodge(width = -0.5),
                 size = 1, 
                 linetype = 1
                 ) +
  scale_color_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                     limits = c("Men",  "Post_menopausal_women","Pre_menopausal_women" )) +
  scale_fill_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                    limits = c("Men",  "Post_menopausal_women","Pre_menopausal_women" )) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower,orupper)  +   
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14)
  ) +
  guides(shape = "none") +
  theme_classic() +
  ylab("Coefficient") 

p

FileName <- paste(FolderName, "/Forest Plot - Ferritin - Logistic Model 2 - Group.png",sep="")
ggsave(FileName, width = 10, height = 7)

```







# Model 3 [Corrected + Donation history]

## Model 3: Men

```{r men m3}
# men
logistic_men_corrected_donhist <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        #Smoking + 
                        #iron_complience + 
                        donation_count_2 + 
                        donation_count +
                        log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort, 
                      data = regression_data_men, 
                      family = bernoulli(),
                      #file = file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_men_corrected_donhist)

```
## Model 3 [Corrected + Donation History]: Premenopausal Women

```{r  pre m3}
# premenopausal women
logistic_pre_corrected_donhist <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        #Smoking + 
                        #iron_complience + 
                        donation_count_2 + 
                        donation_count +
                        log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort, # +
                        #Menstruation +
                        #PreviousChildbirth, 
                      data = regression_data_pre,
                      family = bernoulli(),
                      #file=file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_pre_corrected_donhist)


```
## Model 3 [Corrected + Donation History]: Postmenopausal Women
```{r  post m3}
# postmenopausal women
logistic_post_corrected_donhist <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        #Smoking + 
                        #iron_complience + 
                        donation_count_2 + 
                        donation_count +
                        log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort, #+
                        #PreviousChildbirth,
                     data = regression_data_post, 
                     family = bernoulli(),
                     #file=file, 
                     cores = 4,
                     iter= 10000
)

summary(logistic_post_corrected_donhist)
```

#  Model 3: Retrieve CI data
```{r }
# Group: Men
intervals_men_corrected_donhist <- mcmc_intervals_data(
  logistic_men_corrected_donhist,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Men')


intervals_men_corrected_donhist <- intervals_men_corrected_donhist  %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)'))  %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>%
  filter(! (parameter == 'exp(lprior')) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 


# Group: Premenopausal women
intervals_pre_corrected_donhist <- mcmc_intervals_data(
  logistic_pre_corrected_donhist,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Pre_menopausal_women') %>% 
  arrange(m) # how to order the variables?

intervals_pre_corrected_donhist <- intervals_pre_corrected_donhist  %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)'))  %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>%
  filter(! (parameter == 'exp(lprior')) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

# Group: Postmenopausal women
intervals_post_corrected_donhist <- mcmc_intervals_data(
  logistic_post_corrected_donhist,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Post_menopausal_women')

  
  intervals_post_corrected_donhist <- intervals_post_corrected_donhist  %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)'))  %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>%
  filter(! (parameter == 'exp(lprior')) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values))

# Group: Combined
intervals_combined_corrected_donhist <- rbind(intervals_pre_corrected_donhist,intervals_post_corrected_donhist,intervals_men_corrected_donhist) %>% 
  filter(! (parameter == 'b_Intercept' | parameter == 'lp__' | parameter == 'sigma')) %>% 
  mutate(parameter = gsub("b_","",parameter)) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) #%>% 
  #select(parameter)
  #arrange(m)


```
# Model 3: Forest Plot - Men, Premonpausal women, Postmenopausal women
```{r , fig3, fig.height = 10, fig.width = 15 ,warning=FALSE}

orlower <- -1.5
orupper <- 10

# Truncate ORs for plotting
plotdata_combined_corrected_donhist <- intervals_combined_corrected_donhist %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )

#theme_set(bayesplot::theme_default())
pos <- position_nudge(y=as.numeric(as.factor(plotdata_combined_corrected_donhist$Group))/5 -0.1) 

p <- ggplot(plotdata_combined_corrected_donhist, aes(x = m, y = regressor, color = Group)) + 
  geom_point(position = pos,
             size=4) +
  geom_linerange(aes(xmin = ll, xmax = hh), 
                 position = pos,
                 size = 1, 
                 linetype = 1
                 ) +
  scale_color_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                     limits = c("Men",  "Post_menopausal_women","Pre_menopausal_women" )) +
  scale_fill_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                    limits = c("Men",  "Post_menopausal_women","Pre_menopausal_women" )) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower,orupper)  +   
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14)
  ) +
  guides(shape = "none") +
  theme_classic() +
  ylab("Coefficient") 

p

FileName <- paste(FolderName, "/Forest Plot - Ferritin - Logistic Model 3 - Group.png",sep="")
ggsave(FileName, width = 10, height = 7)

```



# Model 4 [Corrected + Donation History + Smoking/Menstruation/OCC use/Childbirth - Cohort NEW]

# Select data for model 4 (M4)
```{r}
regression_data_men_M4 <- regression_data_men 

regression_data_pre_M4 <- regression_data_pre 

regression_data_post_M4 <- regression_data_post 
```


# Model 4: Men
```{r  men m4}
# men

logistic_men_M4 <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        Smoking + 
                        #iron_complience + 
                        donation_count_2 + 
                        donation_count +
                        log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort, # +
                        #Menstruation +
                        #PreviousChildbirth, 
                      data = regression_data_men_M4,
                      family = bernoulli(),
                      #file=file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_men_M4)


```


```{r  pre m4}
# premenopausal women

logistic_pre_M4 <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        Smoking + 
                        #iron_complience + 
                        donation_count_2 + 
                        donation_count +
                        log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort +
                        Menstruation +
                        PreviousChildbirth +
                        HormonalContraception, 
                      data = regression_data_pre_M4,
                      family = bernoulli(),
                      #file=file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_pre_M4)


```


```{r  post m4}
# postmenopausal women

logistic_post_M4 <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        Smoking + 
                        #iron_complience + 
                        donation_count_2 + 
                        donation_count +
                        log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort +
                        Menstruation +
                        PreviousChildbirth +
                        HormonalContraception, 
                      data = regression_data_post_M4,
                      family = bernoulli(),
                      #file=file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_post_M4)


```




#  Model 4: Retrieve CI data
```{r }
# Group: Men
intervals_men_M4 <- mcmc_intervals_data(
  logistic_men_M4,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Men')


intervals_men_M4 <- intervals_men_M4  %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)'))  %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>%
  filter(! (parameter == 'exp(lprior')) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

# Group: Premenopausal women
intervals_pre_M4 <- mcmc_intervals_data(
  logistic_pre_M4,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Pre_menopausal_women') 

intervals_pre_M4 <- intervals_pre_M4  %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)'))  %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>%
  filter(! (parameter == 'exp(lprior')) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

# Group: Postmenopausal women
intervals_post_M4 <- mcmc_intervals_data(
  logistic_post_M4,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Post_menopausal_women')

intervals_post_M4 <- intervals_post_M4  %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)'))  %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>%
  filter(! (parameter == 'exp(lprior')) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

# Group: Combined
intervals_combined_M4 <- rbind(intervals_pre_M4,intervals_post_M4,intervals_men_M4) %>% 
  filter(! (parameter == 'b_Intercept' | parameter == 'lp__' | parameter == 'sigma')) %>% 
  mutate(parameter = gsub("b_","",parameter)) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) #%>% 
  #select(parameter)
  #arrange(m)


```


# Model 4: Forest Plot - Men, Premonpausal women, Postmenopausal women
```{r}

orlower <- -1
orupper <- 10

# Truncate ORs for plotting
plotdata_combined_M4 <- intervals_combined_M4 %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )

#theme_set(bayesplot::theme_default())
pos <- position_nudge(y=as.numeric(as.factor(plotdata_combined_M4$Group))/5 -0.1) 

p <- ggplot(plotdata_combined_M4, aes(x = m, y = parameter, color = Group)) + 
  geom_point(size = 4, position = position_dodge(width = -1)) +
  geom_linerange(aes(xmin = ll, 
                     xmax = hh),
                    size = 1, 
                    linetype = 1, 
                    position = position_dodge(width = -1))  +
  scale_color_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                     limits = c("Men",  "Post_menopausal_women","Pre_menopausal_women" )) +
  scale_fill_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                    limits = c("Men",  "Post_menopausal_women","Pre_menopausal_women" )) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower,orupper)  +   
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14)
  ) +
  guides(shape = "none") +
  theme_classic() +
  ylab("Coefficient") 

p
# 
FileName <- paste(FolderName, "/Forest Plot - Ferritin - Logistic Model 4 - Group.png",sep="")
ggsave(FileName, width = 10, height = 7)

```




# Final Forest: MEN

```{r}
# Create final forest plot dataset
FinalForest_Model1_Men <- intervals_men
FinalForest_Model2_Men <- intervals_men_corrected
FinalForest_Model3_Men <- intervals_men_corrected_donhist
FinalForest_Model4_Men <- intervals_men_M4

FinalForest_Model1_Men <- FinalForest_Model1_Men[!(FinalForest_Model1_Men$parameter == "lprior"), ]
FinalForest_Model1_Men <- FinalForest_Model1_Men %>% 
  mutate(Model = "Model 1")

FinalForest_Model2_Men <- FinalForest_Model2_Men[!(FinalForest_Model2_Men$parameter == "lprior"), ]
FinalForest_Model2_Men <- FinalForest_Model2_Men %>% 
  mutate(Model = "Model 2")

FinalForest_Model3_Men <- FinalForest_Model3_Men[!(FinalForest_Model3_Men$parameter == "lprior"), ]
FinalForest_Model3_Men <- FinalForest_Model3_Men %>% 
  mutate(Model = "Model 3")

FinalForest_Model4_Men <- FinalForest_Model4_Men[!(FinalForest_Model4_Men$parameter == "lprior"), ]
FinalForest_Model4_Men <- FinalForest_Model4_Men %>% 
  mutate(Model = "Model 4")


# Combine datasets
FinalForest_ModelsCombined <- rbind(FinalForest_Model1_Men,FinalForest_Model2_Men,FinalForest_Model3_Men, FinalForest_Model4_Men)
FinalForest_ModelsCombined <- FinalForest_ModelsCombined[!(FinalForest_ModelsCombined$regressor == "exp(lprior"), ]
FinalForest_ModelsCombined$Model <- as.factor(FinalForest_ModelsCombined$Model)
FinalForest_ModelsCombined$parameter <- gsub("CohortGeneralPopulation", "General Population", FinalForest_ModelsCombined$parameter)
FinalForest_ModelsCombined$parameter <- gsub("donation_count_2", "Number of blood donations (quadratic)", FinalForest_ModelsCombined$parameter)
FinalForest_ModelsCombined$parameter <- gsub("donation_count", "Number of blood donations", FinalForest_ModelsCombined$parameter)
FinalForest_ModelsCombined$parameter <- gsub("log_last_donation", "Days since last blood donation (log transformed)", FinalForest_ModelsCombined$parameter)
FinalForest_ModelsCombined$parameter <- gsub("blood_vol", "Blood volume (Nadler)", FinalForest_ModelsCombined$parameter)
FinalForest_ModelsCombined$parameter <- gsub("Smokingyes", "Smoking", FinalForest_ModelsCombined$parameter)
FinalForest_ModelsCombined$parameter <- as.factor(FinalForest_ModelsCombined$parameter)


desired_order <- c("General Population", "Age", "Blood volume (Nadler)", "Number of blood donations", "Number of blood donations (quadratic)", "Days since last blood donation (log transformed)", "Smoking")
FinalForest_ModelsCombined$parameter <- factor(FinalForest_ModelsCombined$parameter, levels = rev(desired_order))


plotdata_combined_M4 <- FinalForest_ModelsCombined %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )

orlower_men <- -1
orupper_men <- 15

p <-  ggplot(plotdata_combined_M4, aes(x = m, y = parameter, color = parameter)) +
  geom_point(size=4) +
  geom_linerange(aes(xmin = ll, xmax = hh),
                 size = 1,
                 linetype = 1
                 ) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower_men,orupper_men)  +
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme_classic() +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        panel.grid.major.y = element_line(color = "gray", linetype = "dotted"),
        panel.grid.major.x = element_blank()
  ) +
  guides(shape = "none") +
  labs(title = "Logistic Models [Men]") +
  xlab("Coefficient") +
  ylab(" ") +
  facet_grid(Model ~ .,scale="free_y")


p


```

# Final Forest: PREMENOPAUSAL WOMEN

```{r}
# Create final forest plot dataset
FinalForest_Model1_Pre <- intervals_pre
FinalForest_Model2_Pre <- intervals_pre_corrected
FinalForest_Model3_Pre <- intervals_pre_corrected_donhist
FinalForest_Model4_Pre <- intervals_pre_M4

FinalForest_Model1_Pre <- FinalForest_Model1_Pre[!(FinalForest_Model1_Pre$parameter == "lprior"), ]
FinalForest_Model1_Pre <- FinalForest_Model1_Pre %>% 
  mutate(Model = "Model 1")

FinalForest_Model2_Pre <- FinalForest_Model2_Pre[!(FinalForest_Model2_Pre$parameter == "lprior"), ]
FinalForest_Model2_Pre <- FinalForest_Model2_Pre %>% 
  mutate(Model = "Model 2")

FinalForest_Model3_Pre <- FinalForest_Model3_Pre[!(FinalForest_Model3_Pre$parameter == "lprior"), ]
FinalForest_Model3_Pre <- FinalForest_Model3_Pre %>% 
  mutate(Model = "Model 3")

FinalForest_Model4_Pre <- FinalForest_Model4_Pre[!(FinalForest_Model4_Pre$parameter == "lprior"), ]
FinalForest_Model4_Pre <- FinalForest_Model4_Pre %>% 
  mutate(Model = "Model 4")


# Combine datasets
FinalForest_ModelsCombined_Pre <- rbind(FinalForest_Model1_Pre,FinalForest_Model2_Pre,FinalForest_Model3_Pre, FinalForest_Model4_Pre)
FinalForest_ModelsCombined_Pre <- FinalForest_ModelsCombined_Pre[!(FinalForest_ModelsCombined_Pre$parameter == "exp(lprior"), ]
FinalForest_ModelsCombined_Pre$Model <- as.factor(FinalForest_ModelsCombined_Pre$Model)
FinalForest_ModelsCombined_Pre$parameter <- gsub("CohortGeneralPopulation", "General Population", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("donation_count_2", "Number of blood donations (quadratic)", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("donation_count", "Number of blood donations", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("log_last_donation", "Days since last blood donation (log transformed)", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("blood_vol", "Blood volume (Nadler)", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("Smokingyes", "Smoking", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("MenstruationYes", "Menstruation", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("PreviousChildbirthyes", "Childbirth in the past", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("HormonalContraceptiones", "Hormonal contraception", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- as.factor(FinalForest_ModelsCombined_Pre$parameter)



desired_order <- c("General Population", "Age", "Blood volume (Nadler)", "Number of blood donations", "Number of blood donations (quadratic)", "Days since last blood donation (log transformed)", "Smoking", "Menstruation", "Childbirth", "Oral contraception")

plotdata_combined_M4_Pre <- FinalForest_ModelsCombined_Pre %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )

orlower_men <- -1
orupper_men <- 3

p <-  ggplot(plotdata_combined_M4_Pre, aes(x = m, y = parameter, color = parameter)) +
  geom_point(size=4) +
  geom_linerange(aes(xmin = ll, xmax = hh),
                 size = 1,
                 linetype = 1
                 ) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower_men,orupper_men)  +
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme_classic() +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        panel.grid.major.y = element_line(color = "gray", linetype = "dotted"),
        panel.grid.major.x = element_blank()
  ) +
  guides(shape = "none") +
  labs(title = "Logistic Models [Women - Premenopausal]") +
  xlab("Coefficient") +
  ylab(" ") +
  facet_grid(Model ~ .,scale="free_y")

p
# 
FileName <- paste(FolderName, "/Forest Plot - Logistic - Final - Premenopausal.png",sep="")
ggsave(FileName, width = 15, height = 10)


```


# Final Forest: POSTMENOPAUSAL WOMEN

```{r}
# Create final forest plot dataset
FinalForest_Model1_Post <- intervals_post
FinalForest_Model2_Post <- intervals_post_corrected
FinalForest_Model3_Post <- intervals_post_corrected_donhist
FinalForest_Model4_Post <- intervals_post_M4

FinalForest_Model1_Post <- FinalForest_Model1_Post[!(FinalForest_Model1_Post$parameter == "lprior"), ]
FinalForest_Model1_Post <- FinalForest_Model1_Post %>% 
  mutate(Model = "Model 1")

FinalForest_Model2_Post <- FinalForest_Model2_Post[!(FinalForest_Model2_Post$parameter == "lprior"), ]
FinalForest_Model2_Post <- FinalForest_Model2_Post %>% 
  mutate(Model = "Model 2")

FinalForest_Model3_Post <- FinalForest_Model3_Post[!(FinalForest_Model3_Post$parameter == "lprior"), ]
FinalForest_Model3_Post <- FinalForest_Model3_Post %>% 
  mutate(Model = "Model 3")

FinalForest_Model4_Post <- FinalForest_Model4_Post[!(FinalForest_Model4_Post$parameter == "lprior"), ]
FinalForest_Model4_Post <- FinalForest_Model4_Post %>% 
  mutate(Model = "Model 4")


# Combine datasets
FinalForest_ModelsCombined_Post <- rbind(FinalForest_Model1_Post,FinalForest_Model2_Post,FinalForest_Model3_Post, FinalForest_Model4_Post)
FinalForest_ModelsCombined_Post <- FinalForest_ModelsCombined_Post[!(FinalForest_ModelsCombined_Post$parameter == "exp(lprior"), ]
FinalForest_ModelsCombined_Post$Model <- as.factor(FinalForest_ModelsCombined_Post$Model)
FinalForest_ModelsCombined_Post$parameter <- gsub("CohortGeneralPopulation", "General Population", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("donation_count_2", "Number of blood donations (quadratic)", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("donation_count", "Number of blood donations", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("log_last_donation", "Days since last blood donation (log transformed)", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("blood_vol", "Blood volume (Nadler)", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("Smokingyes", "Smoking", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("MenstruationYes", "Menstruation", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("PreviousChildbirthyes", "Childbirth in the past", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("HormonalContraceptionyes", "Hormonal contraception", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- as.factor(FinalForest_ModelsCombined_Post$parameter)

desired_order <- c("General Population", "Age", "Blood volume (Nadler)", "Number of blood donations", "Number of blood donations (quadratic)", "Days since last blood donation (log transformed)", "Smoking", "Menstruation", "Childbirth in the past", "Hormonal contraception")
FinalForest_ModelsCombined_Post$parameter <- factor(FinalForest_ModelsCombined_Post$parameter, levels = rev(desired_order))


plotdata_combined_M4_Post <- FinalForest_ModelsCombined_Post %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )


orlower <- -1
orupper <- 20

p <-  ggplot(plotdata_combined_M4_Post, aes(x = m, y = parameter, color = parameter)) +
  geom_point(size=4) +
  geom_linerange(aes(xmin = ll, xmax = hh),
                 size = 1,
                 linetype = 1
                 ) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower,orupper)  +
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme_classic() +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        panel.grid.major.y = element_line(color = "gray", linetype = "dotted"),
        panel.grid.major.x = element_blank()
  ) +
  guides(shape = "none") +
  labs(title = "Logistic Models [Women - Postmenopausal]") +
  xlab("Coefficient") +
  ylab(" ") +
  facet_grid(Model ~ .,scale="free_y")

p

FileName <- paste(FolderName, "/Forest Plot - Logistic - Final - Postmenopausal.png",sep="")
ggsave(FileName, width = 15, height = 10)


```


# Final forest all groups dataset
```{r, fig4, width = 15, height = 10}
FinalForest_ModelsCombined_2 <- FinalForest_ModelsCombined
FinalForest_ModelsCombined_2$Group <- "Men"
FinalForest_ModelsCombined_Pre_2 <- FinalForest_ModelsCombined_Pre
FinalForest_ModelsCombined_Pre_2$Group <- "Women Premenopausal"
FinalForest_ModelsCombined_Post_2 <- FinalForest_ModelsCombined_Post
FinalForest_ModelsCombined_Post_2$Group <- "Women Postmenopausal"

FinalForest_ModelsCombined_All_2 <- rbind(FinalForest_ModelsCombined_2,FinalForest_ModelsCombined_Pre_2,FinalForest_ModelsCombined_Post_2)
FinalForest_ModelsCombined_All_2$Group <- factor(FinalForest_ModelsCombined_All_2$Group, levels = c("Men", "Women Premenopausal", "Women Postmenopausal"))
FinalForest_ModelsCombined_All_2$parameter <- gsub("Blood_volume", "Blood volume", FinalForest_ModelsCombined_All_2$parameter)


FinalForest_ModelsCombined_All_2$parameter <- gsub("MenstruationYes", "Menstruation", FinalForest_ModelsCombined_All_2$parameter)
FinalForest_ModelsCombined_All_2$parameter <- gsub("PreviousChildbirthyes", "Childbirth in the past", FinalForest_ModelsCombined_All_2$parameter)
FinalForest_ModelsCombined_All_2$parameter <- gsub("HormonalContraceptionyes", "Hormonal contraception", FinalForest_ModelsCombined_All_2$parameter)





desired_order_parameter <- c("General Population", "Age", "Blood volume (Nadler)", "Number of blood donations", "Number of blood donations (quadratic)", "Days since last blood donation (log transformed)", "Smoking", "Menstruation", "Childbirth in the past", "Hormonal contraception")
FinalForest_ModelsCombined_All_2$parameter <- factor(FinalForest_ModelsCombined_All_2$parameter, levels = rev(desired_order_parameter))


plotdata_ModelsCombined_All_2 <- FinalForest_ModelsCombined_All_2 %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )

orlower_men <- -1
orupper_men <- 20

p <- ggplot(plotdata_ModelsCombined_All_2, aes(x = m, y = parameter, color = parameter)) +
  geom_point(size=4) +
  geom_linerange(aes(xmin = ll, xmax = hh),
                 size = 1,
                 linetype = 1
                 ) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower_men,orupper_men)  +
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme_classic() +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        panel.grid.major.y = element_line(color = "gray", linetype = "dotted"),
        panel.grid.major.x = element_blank(),
        legend.position = "none"
  ) +
  guides(shape = "none") +
  labs(title = "Logistic Models") +
  xlab("Coefficient") +
  ylab(" ") +
  facet_grid(Model ~ Group,scale="free_y")



p
# 
FileName <- paste(FolderName, "/Forest Plot - Logistic - Final - All Groups.png",sep="")
ggsave(FileName, width = 15, height = 10)




```



# Save Logistic regression Summary Data Dataset

```{r}
FinalForest_FIN_log <- FinalForest_ModelsCombined_All_2 %>% 
  add_column(Country = "FIN")

save(FinalForest_FIN_log, file = "~/proj/iron_measurement_comparisons/results/genpop_combined/low_ferritin_model_bayes/FinalForest_FIN_log.rdata")

saveRDS(FinalForest_ModelsCombined_All_2, file.path("~/proj/iron_measurement_comparisons/results/genpop_combined/low_ferritin_model_bayes/", "3. Cohort Comparison - Summary Data - Linear Regression.rds"))


```








BAYESIAN LOGISTIC REGRESSION WITH SCALED CO-VARIATES

## Data
```{r}
# men
## menstruation and childbirth variables removed 
regression_data_men_scaled <- all_cohorts %>% 
  filter(Group == "Men") %>% 
  dplyr::select(ID, log_ferritin, Age, log_CRP, blood_vol, donation_count, log_last_donation, Cohort, iron_deficiency, Smoking, Group, Weight) %>%
  mutate(Age = scale(Age, scale = TRUE)[,1],
         log_CRP = scale(log_CRP, scale = TRUE)[,1],
         donation_count = scale(donation_count, scale = TRUE)[,1],
         log_last_donation = scale(log_last_donation, scale = TRUE)[,1],
         blood_vol = scale(blood_vol, scale = TRUE)[,1],
         donation_count_2 = donation_count^2,
         donation_count_2 = scale(donation_count_2, scale = TRUE)[,1]
         )
         

# premenopausal women
regression_data_pre_scaled <- all_cohorts %>% 
  filter(Group == "Pre_menopause_women") %>% 
  dplyr::select(ID, log_ferritin, Age, log_CRP, blood_vol, donation_count, log_last_donation, Cohort, iron_deficiency, Smoking, Group, Weight, HormonalContraception, PreviousChildbirth, Menstruation) %>%
  mutate(Age = scale(Age, scale = TRUE)[,1],
         log_CRP = scale(log_CRP, scale = TRUE)[,1],
         donation_count = scale(donation_count, scale = TRUE)[,1],
         log_last_donation = scale(log_last_donation, scale = TRUE)[,1],
         blood_vol = scale(blood_vol, scale = TRUE)[,1],
         donation_count_2 = donation_count^2,
         donation_count_2 = scale(donation_count_2, scale = TRUE)[,1]
         ) 

# postmenopausal womem 
## menstruation variable removed 
regression_data_post_scaled <- all_cohorts %>% 
  filter(Group == "Post_menopause_women") %>% 
  dplyr::select(ID, log_ferritin, Age, log_CRP, blood_vol, donation_count, log_last_donation, Cohort, iron_deficiency, Smoking, Group, Weight, HormonalContraception, PreviousChildbirth, Menstruation) %>%
  mutate(Age = scale(Age, scale = TRUE)[,1],
         log_CRP = scale(log_CRP, scale = TRUE)[,1],
         donation_count = scale(donation_count, scale = TRUE)[,1],
         log_last_donation = scale(log_last_donation, scale = TRUE)[,1],
         blood_vol = scale(blood_vol, scale = TRUE)[,1],
         donation_count_2 = donation_count^2,
         donation_count_2 = scale(donation_count_2, scale = TRUE)[,1]
  )


```

# Model 1 [Crude Model]


# Model 1: Men
```{r men m1 scaled}
#men
logistic_men_b_scaled <- brm(iron_deficiency ~ 
                        #Age + 
                        #blood_volume + 
                        #Smoking + 
                        # donation_count_2 + 
                        # donation_count +
                        # log_last_donation + 
                        #log_CRP +
                        #Weight +
                        Cohort, 
                      data = regression_data_men_scaled, 
                      family = bernoulli(),
                      #file=file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_men_b_scaled)



```


# Model 1: Premenopausal Women
```{r pre m1 scaled}

logistic_pre_b_scaled <- brm(iron_deficiency ~ 
                        #Age + 
                        #blood_volume + 
                        #Smoking + 
                        # donation_count_2 + 
                        # donation_count +
                        # log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort, #+
                        #Menstruation +
                        #PreviousChildbirth, 
                      data = regression_data_pre_scaled,
                      family = bernoulli(),
                      #file=file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_pre_b_scaled)



```


# Model 1: Postmenopausal Women
```{r post m1 scaled}

logistic_post_b_scaled <- brm(iron_deficiency ~
                        #Age + 
                        #blood_volume + 
                        # donation_count_2 +
                        # donation_count +
                        # log_last_donation + 
                        #Weight +
                        Cohort,
                        #log_CRP +
                        #Smoking + 
                        #RedMeat_n +
                        #+
                        #Menstruation +   
                        #PreviousChildbirth, 
                     data = regression_data_post_scaled, 
                     family = bernoulli(),
                     #file=file, 
                     cores = 4,
                     iter= 10000
)

summary(logistic_post_b_scaled)



```








## Diagnostics

### Trace

```{r , message=FALSE}
mcmc_plot(logistic_men_b_scaled, type = "trace")
```

```{r , message=FALSE}
mcmc_plot(logistic_pre_b_scaled, type = "trace")
```

```{r , message=FALSE}
mcmc_plot(logistic_post_b_scaled, type = "trace")
```

Chains do not seem to get stuck.

### ACF bars

```{r}
mcmc_plot(logistic_men_b_scaled, type = "acf_bar")
```


```{r}
mcmc_plot(logistic_pre_b_scaled, type = "acf_bar")
```


```{r}
mcmc_plot(logistic_post_b_scaled, type = "acf_bar")
```

Autocorrelation drops nicely.

### Rhat

```{r}
mcmc_plot(logistic_men_b_scaled, type = "rhat")
```

```{r}
mcmc_plot(logistic_pre_b_scaled, type = "rhat")
```

```{r}
mcmc_plot(logistic_post_b_scaled, type = "rhat")
```
https://mc-stan.org/rstan/reference/Rhat.html


"The Rhat function produces R-hat convergence diagnostic, which compares the between- and within-chain estimates for model parameters and other univariate quantities of interest. If chains have not mixed well (ie, the between- and within-chain estimates don't agree), R-hat is larger than 1. We recommend running at least four chains by default and only using the sample if R-hat is less than 1.05."

# Distributions


```{r , message=FALSE, warning=FALSE}
p <- mcmc_plot(logistic_men_b_scaled, 
         type = "areas",
         prob = 0.95,
         transformations = "exp") +
  geom_vline(xintercept = 1, color = "grey")

p <- p + scale_x_log10()
#Ferritin needs to be removed
# plot.OR<-  tidy_logit_pre_menop_women %>%filter(!term=="log_Ferritin_beginning") %>%  dplyr::select(OR) %>% mutate(y=rev(2:(nrow(tidy_logit_pre_menop_women)))) %>% bind_rows(tibble(OR=NA,y=1 ))
# p <- p + geom_point(aes(y=y,x=OR),data=plot.OR, col="red")
p
```


```{r , message=FALSE, warning=FALSE}
p <- mcmc_plot(logistic_pre_b_scaled, 
         type = "areas",
         prob=0.95,
         prob_est="median",
         transformations = "exp"
        ) +
  geom_vline(xintercept = 1, color = "grey")

p <- p + scale_x_log10()
p
```


```{r , message=FALSE, warning=FALSE}
p <- mcmc_plot(logistic_post_b_scaled, 
         type = "areas",
         prob = 0.95,
         transformations = "exp") +
  geom_vline(xintercept = 1, color = "grey")

p <- p + scale_x_log10()
#Ferritin needs to be removed
# plot.OR<-  tidy_logit_pre_menop_women %>%filter(!term=="log_Ferritin_beginning") %>%  dplyr::select(OR) %>% mutate(y=rev(2:(nrow(tidy_logit_pre_menop_women)))) %>% bind_rows(tibble(OR=NA,y=1 ))
# p <- p + geom_point(aes(y=y,x=OR),data=plot.OR, col="red")
p
```

### All groups together

```{r}
regressor_values <- c(
    "Age" = "Age (5 years)",
    "CohortGeneralPopulation" = "General population", 
   "log_CRP" = "CRP (log transformed)",
   "blood_vol" = "Blood volume (Nadler)",
   "log_last_donation" = "Days since last donation (log, /log(2))",
   "Smokingyes" ="Smoking",
    "iron_complience" = "Iron supplementation",
    "donation_count_2" = "Donations in last 2 years (quadratic)",
    "Menstruationregular_period" = "Regular periods",
    "Weight" = "Weight (10 kg)",
    "PreviousChildbirth" = "Given birth in the past",
    "donation_count" = "Donations in last 2 years"
)
```



# Model 1: Intervals
```{r }

intervals_men <- mcmc_intervals_data(
  logistic_men_b_scaled,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Men')

 intervals_men <- intervals_men%>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  filter(! (parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

intervals_pre <- mcmc_intervals_data(
  logistic_pre_b_scaled,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Pre_menopause_women') %>% 
  arrange(m) # how to order the variables?

 intervals_pre <- intervals_pre%>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  filter(! (parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

intervals_post <- mcmc_intervals_data(
  logistic_post_b_scaled,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Post_menopause_women')

 intervals_post <- intervals_post%>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  filter(! (parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

intervals_combined <- rbind(intervals_pre,intervals_post,intervals_men) %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

```

# Model 1: Forest plot - Men, Premonpausal women, Postmenopausal women
```{r , fig5, fig.height = 10, fig.width = 15 ,warning=FALSE}

orlower <- -0.5
orupper <- 2.5

# Truncate ORs for plotting
plotdata <- intervals_combined %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )

#theme_set(bayesplot::theme_default())
pos <- position_nudge(y=as.numeric(as.factor(plotdata$Group))/5 -0.1) 

p <- ggplot(plotdata, aes(x = m, y = regressor, color = Group)) + 
  geom_point(position = position_dodge(width = -0.4),
             size=4) +
  geom_linerange(aes(xmin = ll, xmax = hh), 
                 position = position_dodge(width = -0.4),
                 size = 1, 
                 linetype = 1
                 ) +
  scale_color_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                     limits = c("Men",  "Post_menopause_women","Pre_menopause_women" )) +
  scale_fill_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                    limits = c("Men",  "Post_menopause_women","Pre_menopause_women" )) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower,orupper)  +   
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14)
  ) +
  guides(shape = "none") +
  theme_classic() +
  ylab("Coefficient") 

p

FileName <- paste(FolderName, "/Forest Plot - Logistic - Model 1 scaled.png",sep="")
ggsave(FileName, width = 10, height = 7)

```



# Model 2 [Corrected]

## Model 2 [Corrected]: Men

```{r men m2 scaled}
#men
logistic_men_corrected_scaled <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        #Smoking + 
                        #iron_complience + 
                        #donation_count_2 + 
                        #donation_count +
                        #log_last_donation + 
                        #log_CRP +
                        #Region + 
                        # Weight +
                        #RedMeat_n +
                        Cohort, 
                      data = regression_data_men_scaled, 
                      family = bernoulli(),
                      #file = file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_men_corrected_scaled)

```

## Model 2 [Corrected]: Premenopausal Women

```{r  pre m2 scaled}
# premenopausal women
#file <- "../results/bmodels/version3/linear_pre_b"

logistic_pre_corrected_scaled <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        #Smoking + 
                        #iron_complience + 
                        #donation_count_2 + 
                        #donation_count +
                        #log_last_donation + 
                        #log_CRP +
                        #Region + 
                        # Weight +
                        #RedMeat_n +
                        Cohort, # +
                        #Menstruation +
                        #PreviousChildbirth, 
                      data = regression_data_pre_scaled,
                      family = bernoulli(),
                      #file=file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_pre_corrected_scaled)


```

## Model 2 [Corrected]: Postmenopausal Women

```{r  post m2 scaled}
# postmenopausal women
#file <- "../results/bmodels/version3/linear_post_b"

logistic_post_corrected_scaled <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        #Smoking + 
                        #iron_complience + 
                        #donation_count_2 + 
                        #donation_count +
                        #log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort, #+
                        #PreviousChildbirth,
                     data = regression_data_post_scaled, 
                     family = bernoulli(),
                     #file=file, 
                     cores = 4,
                     iter= 10000
)

summary(logistic_post_corrected_scaled)
```

#  Model 2: Retrieve CI data
```{r }
# Group: Men
intervals_men_corrected <- mcmc_intervals_data(
  logistic_men_corrected_scaled,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
) %>% mutate(Group='Men')

 intervals_men_corrected <- intervals_men_corrected%>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  filter(! (parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

# Group: Premenopausal women
intervals_pre_corrected <- mcmc_intervals_data(
  logistic_pre_corrected_scaled,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  )  %>% mutate(Group='Pre_menopausal_women')

 intervals_pre_corrected <- intervals_pre_corrected%>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  filter(! (parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

# Group: Postmenopausal women
intervals_post_corrected <- mcmc_intervals_data(
  logistic_post_corrected_scaled,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Post_menopausal_women')

 intervals_post_corrected <- intervals_post_corrected%>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  filter(! (parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

# Group: Combined
intervals_combined_corrected <- rbind(intervals_pre_corrected,intervals_post_corrected,intervals_men_corrected) %>% 
  filter(! (parameter == 'b_Intercept' | parameter == 'lp__'  | parameter == 'lprior' | parameter == 'sigma')) %>% 
  mutate(parameter = gsub("b_","",parameter)) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values))

intervals_combined_corrected <- rbind(intervals_pre_corrected,intervals_post_corrected,intervals_men_corrected) %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)' | parameter == 'exp(lprior')) %>%
  filter(! (parameter == 'exp(lprior')) %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

```

# Model 2: Forest plot  [Men, Premonpausal women, Postmenopausal women]
```{r , fig6, fig.height = 10, fig.width = 15 ,warning=FALSE}

orlower <- -0.5
orupper <- 2.5

# Truncate ORs for plotting
plotdata_combined_corrected <- intervals_combined_corrected %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )

#theme_set(bayesplot::theme_default())
pos <- position_nudge(y=as.numeric(as.factor(plotdata$Group))/5 -0.1) 

p <- ggplot(plotdata_combined_corrected, aes(x = m, y = regressor, color = Group)) + 
  geom_point(position = position_dodge(width = -0.5),
             size=4) +
  geom_linerange(aes(xmin = ll, xmax = hh), 
                 position = position_dodge(width = -0.5),
                 size = 1, 
                 linetype = 1
                 ) +
  scale_color_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                     limits = c("Men",  "Post_menopausal_women","Pre_menopausal_women" )) +
  scale_fill_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                    limits = c("Men",  "Post_menopausal_women","Pre_menopausal_women" )) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower,orupper)  +   
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14)
  ) +
  guides(shape = "none") +
  theme_classic() +
  ylab("Coefficient") 

p

FileName <- paste(FolderName, "/Forest Plot - Ferritin - Logistic Model 2 - Group scaled.png",sep="")
ggsave(FileName, width = 10, height = 7)

```







# Model 3 [Corrected + Donation history]

## Model 3: Men

```{r men m3 scaled}
# men
logistic_men_corrected_donhist_scaled <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        #Smoking + 
                        #iron_complience + 
                        donation_count_2 + 
                        donation_count +
                        log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort, 
                      data = regression_data_men_scaled, 
                      family = bernoulli(),
                      #file = file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_men_corrected_donhist_scaled)

```
## Model 3 [Corrected + Donation History]: Premenopausal Women

```{r  pre m3 scaled}
# premenopausal women
logistic_pre_corrected_donhist_scaled <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        #Smoking + 
                        #iron_complience + 
                        donation_count_2 + 
                        donation_count +
                        log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort, # +
                        #Menstruation +
                        #PreviousChildbirth, 
                      data = regression_data_pre_scaled,
                      family = bernoulli(),
                      #file=file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_pre_corrected_donhist_scaled)


```
## Model 3 [Corrected + Donation History]: Postmenopausal Women
```{r  post m3 scaled}
# postmenopausal women
logistic_post_corrected_donhist_scaled <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        #Smoking + 
                        #iron_complience + 
                        donation_count_2 + 
                        donation_count +
                        log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort, #+
                        #PreviousChildbirth,
                     data = regression_data_post_scaled, 
                     family = bernoulli(),
                     #file=file, 
                     cores = 4,
                     iter= 10000
)

summary(logistic_post_corrected_donhist_scaled)
```

#  Model 3: Retrieve CI data
```{r }
# Group: Men
intervals_men_corrected_donhist <- mcmc_intervals_data(
  logistic_men_corrected_donhist_scaled,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Men')


intervals_men_corrected_donhist <- intervals_men_corrected_donhist  %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)'))  %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>%
  filter(! (parameter == 'exp(lprior')) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 


# Group: Premenopausal women
intervals_pre_corrected_donhist <- mcmc_intervals_data(
  logistic_pre_corrected_donhist_scaled,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Pre_menopausal_women') %>% 
  arrange(m) # how to order the variables?

intervals_pre_corrected_donhist <- intervals_pre_corrected_donhist  %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)'))  %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>%
  filter(! (parameter == 'exp(lprior')) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

# Group: Postmenopausal women
intervals_post_corrected_donhist <- mcmc_intervals_data(
  logistic_post_corrected_donhist_scaled,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Post_menopausal_women')

  
  intervals_post_corrected_donhist <- intervals_post_corrected_donhist  %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)'))  %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>%
  filter(! (parameter == 'exp(lprior')) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values))

# Group: Combined
intervals_combined_corrected_donhist <- rbind(intervals_pre_corrected_donhist,intervals_post_corrected_donhist,intervals_men_corrected_donhist) %>% 
  filter(! (parameter == 'b_Intercept' | parameter == 'lp__' | parameter == 'sigma')) %>% 
  mutate(parameter = gsub("b_","",parameter)) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) #%>% 
  #select(parameter)
  #arrange(m)


```
# Model 3: Forest Plot - Men, Premonpausal women, Postmenopausal women
```{r , fig7, fig.height = 10, fig.width = 15 ,warning=FALSE}

orlower <- -1.5
orupper <- 10

# Truncate ORs for plotting
plotdata_combined_corrected_donhist <- intervals_combined_corrected_donhist %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )

#theme_set(bayesplot::theme_default())
pos <- position_nudge(y=as.numeric(as.factor(plotdata_combined_corrected_donhist$Group))/5 -0.1) 

p <- ggplot(plotdata_combined_corrected_donhist, aes(x = m, y = regressor, color = Group)) + 
  geom_point(position = pos,
             size=4) +
  geom_linerange(aes(xmin = ll, xmax = hh), 
                 position = pos,
                 size = 1, 
                 linetype = 1
                 ) +
  scale_color_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                     limits = c("Men",  "Post_menopausal_women","Pre_menopausal_women" )) +
  scale_fill_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                    limits = c("Men",  "Post_menopausal_women","Pre_menopausal_women" )) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower,orupper)  +   
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14)
  ) +
  guides(shape = "none") +
  theme_classic() +
  ylab("Coefficient") 

p

FileName <- paste(FolderName, "/Forest Plot - Ferritin - Logistic Model 3 - Group scaled.png",sep="")
ggsave(FileName, width = 10, height = 7)

```



# Model 4 [Corrected + Donation History + Smoking/Menstruation/OCC use/Childbirth - Cohort NEW]

# Select data for model 4 (M4)
```{r}
regression_data_men_M4_scaled <- regression_data_men_scaled 

regression_data_pre_M4_scaled <- regression_data_pre_scaled 

regression_data_post_M4_scaled <- regression_data_post_scaled 
```


# Model 4: Men
```{r  men m4 scaled}
# men

logistic_men_M4_scaled <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        Smoking + 
                        #iron_complience + 
                        donation_count_2 + 
                        donation_count +
                        log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort, # +
                        #Menstruation +
                        #PreviousChildbirth, 
                      data = regression_data_men_M4_scaled,
                      family = bernoulli(),
                      #file=file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_men_M4_scaled)


```


```{r  pre m4 scaled}
# premenopausal women

logistic_pre_M4_scaled <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        Smoking + 
                        #iron_complience + 
                        donation_count_2 + 
                        donation_count +
                        log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort +
                        Menstruation +
                        PreviousChildbirth +
                        HormonalContraception, 
                      data = regression_data_pre_M4_scaled,
                      family = bernoulli(),
                      #file=file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_pre_M4_scaled)


```


```{r  post m4 scaled}
# postmenopausal women

logistic_post_M4_scaled <- brm(iron_deficiency ~ 
                        Age + 
                        blood_vol + 
                        Smoking + 
                        #iron_complience + 
                        donation_count_2 + 
                        donation_count +
                        log_last_donation + 
                        #log_CRP +
                        #Region + 
                        #Weight +
                        #RedMeat_n +
                        Cohort +
                        Menstruation +
                        PreviousChildbirth +
                        HormonalContraception, 
                      data = regression_data_post_M4_scaled,
                      family = bernoulli(),
                      #file=file, 
                      cores = 4,
                      iter= 10000
)

summary(logistic_post_M4_scaled)


```




#  Model 4: Retrieve CI data
```{r }
# Group: Men
intervals_men_M4 <- mcmc_intervals_data(
  logistic_men_M4_scaled,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Men')


intervals_men_M4 <- intervals_men_M4  %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)'))  %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>%
  filter(! (parameter == 'exp(lprior')) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

# Group: Premenopausal women
intervals_pre_M4 <- mcmc_intervals_data(
  logistic_pre_M4_scaled,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Pre_menopausal_women') 

intervals_pre_M4 <- intervals_pre_M4  %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)'))  %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>%
  filter(! (parameter == 'exp(lprior')) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

# Group: Postmenopausal women
intervals_post_M4 <- mcmc_intervals_data(
  logistic_post_M4_scaled,
  prob=0.9,
  prob_outer = 0.95,
  point_est = "median",
  transformations = "exp"
  ) %>% mutate(Group='Post_menopausal_women')

intervals_post_M4 <- intervals_post_M4  %>% 
  filter(! (parameter == 'exp(b_Intercept)' | parameter == 'exp(lp__)'))  %>%
  mutate(parameter = gsub("\\)","",gsub("exp\\(b_","",parameter))) %>%
  filter(! (parameter == 'exp(lprior')) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) 

# Group: Combined
intervals_combined_M4 <- rbind(intervals_pre_M4,intervals_post_M4,intervals_men_M4) %>% 
  filter(! (parameter == 'b_Intercept' | parameter == 'lp__' | parameter == 'sigma')) %>% 
  mutate(parameter = gsub("b_","",parameter)) %>% 
  mutate(regressor = plyr::revalue(parameter, regressor_values)) #%>% 
  #select(parameter)
  #arrange(m)


```


# Model 4: Forest Plot - Men, Premonpausal women, Postmenopausal women
```{r}

orlower <- -1
orupper <- 10

# Truncate ORs for plotting
plotdata_combined_M4 <- intervals_combined_M4 %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )

#theme_set(bayesplot::theme_default())
pos <- position_nudge(y=as.numeric(as.factor(plotdata_combined_M4$Group))/5 -0.1) 

p <- ggplot(plotdata_combined_M4, aes(x = m, y = parameter, color = Group)) + 
  geom_point(size = 4, position = position_dodge(width = -1)) +
  geom_linerange(aes(xmin = ll, 
                     xmax = hh),
                    size = 1, 
                    linetype = 1, 
                    position = position_dodge(width = -1))  +
  scale_color_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                     limits = c("Men",  "Post_menopausal_women","Pre_menopausal_women" )) +
  scale_fill_manual(values = c(  "#0065BD", "#BA0303", "#C50084" ),
                    limits = c("Men",  "Post_menopausal_women","Pre_menopausal_women" )) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower,orupper)  +   
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14)
  ) +
  guides(shape = "none") +
  theme_classic() +
  ylab("Coefficient") 

p

FileName <- paste(FolderName, "/Forest Plot - Ferritin - Logistic Model 4 - Group scaled.png",sep="")
ggsave(FileName, width = 10, height = 7)

```




# Final Forest: MEN

```{r}
# Create final forest plot dataset
FinalForest_Model1_Men <- intervals_men
FinalForest_Model2_Men <- intervals_men_corrected
FinalForest_Model3_Men <- intervals_men_corrected_donhist
FinalForest_Model4_Men <- intervals_men_M4

FinalForest_Model1_Men <- FinalForest_Model1_Men[!(FinalForest_Model1_Men$parameter == "lprior"), ]
FinalForest_Model1_Men <- FinalForest_Model1_Men %>% 
  mutate(Model = "Model 1")

FinalForest_Model2_Men <- FinalForest_Model2_Men[!(FinalForest_Model2_Men$parameter == "lprior"), ]
FinalForest_Model2_Men <- FinalForest_Model2_Men %>% 
  mutate(Model = "Model 2")

FinalForest_Model3_Men <- FinalForest_Model3_Men[!(FinalForest_Model3_Men$parameter == "lprior"), ]
FinalForest_Model3_Men <- FinalForest_Model3_Men %>% 
  mutate(Model = "Model 3")

FinalForest_Model4_Men <- FinalForest_Model4_Men[!(FinalForest_Model4_Men$parameter == "lprior"), ]
FinalForest_Model4_Men <- FinalForest_Model4_Men %>% 
  mutate(Model = "Model 4")


# Combine datasets
FinalForest_ModelsCombined <- rbind(FinalForest_Model1_Men,FinalForest_Model2_Men,FinalForest_Model3_Men, FinalForest_Model4_Men)
FinalForest_ModelsCombined <- FinalForest_ModelsCombined[!(FinalForest_ModelsCombined$regressor == "exp(lprior"), ]
FinalForest_ModelsCombined$Model <- as.factor(FinalForest_ModelsCombined$Model)
FinalForest_ModelsCombined$parameter <- gsub("CohortGeneralPopulation", "General Population", FinalForest_ModelsCombined$parameter)
FinalForest_ModelsCombined$parameter <- gsub("donation_count_2", "Number of blood donations (quadratic)", FinalForest_ModelsCombined$parameter)
FinalForest_ModelsCombined$parameter <- gsub("donation_count", "Number of blood donations", FinalForest_ModelsCombined$parameter)
FinalForest_ModelsCombined$parameter <- gsub("log_last_donation", "Days since last blood donation (log transformed)", FinalForest_ModelsCombined$parameter)
FinalForest_ModelsCombined$parameter <- gsub("blood_vol", "Blood volume (Nadler)", FinalForest_ModelsCombined$parameter)
FinalForest_ModelsCombined$parameter <- gsub("Smokingyes", "Smoking", FinalForest_ModelsCombined$parameter)
FinalForest_ModelsCombined$parameter <- as.factor(FinalForest_ModelsCombined$parameter)


desired_order <- c("General Population", "Age", "Blood volume (Nadler)", "Number of blood donations", "Number of blood donations (quadratic)", "Days since last blood donation (log transformed)", "Smoking")
FinalForest_ModelsCombined$parameter <- factor(FinalForest_ModelsCombined$parameter, levels = rev(desired_order))


plotdata_combined_M4 <- FinalForest_ModelsCombined %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )

orlower_men <- -1
orupper_men <- 15

p <-  ggplot(plotdata_combined_M4, aes(x = m, y = parameter, color = parameter)) +
  geom_point(size=4) +
  geom_linerange(aes(xmin = ll, xmax = hh),
                 size = 1,
                 linetype = 1
                 ) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower_men,orupper_men)  +
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme_classic() +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        panel.grid.major.y = element_line(color = "gray", linetype = "dotted"),
        panel.grid.major.x = element_blank()
  ) +
  guides(shape = "none") +
  labs(title = "Logistic Models [Men], scaled") +
  xlab("Coefficient") +
  ylab(" ") +
  facet_grid(Model ~ .,scale="free_y")


p

FileName <- paste(FolderName, "/Forest Plot - Logistic - Final - Men scaled.png",sep="")
ggsave(FileName, width = 15, height = 10)


```

# Final Forest: PREMENOPAUSAL WOMEN

```{r}
# Create final forest plot dataset
FinalForest_Model1_Pre <- intervals_pre
FinalForest_Model2_Pre <- intervals_pre_corrected
FinalForest_Model3_Pre <- intervals_pre_corrected_donhist
FinalForest_Model4_Pre <- intervals_pre_M4

FinalForest_Model1_Pre <- FinalForest_Model1_Pre[!(FinalForest_Model1_Pre$parameter == "lprior"), ]
FinalForest_Model1_Pre <- FinalForest_Model1_Pre %>% 
  mutate(Model = "Model 1")

FinalForest_Model2_Pre <- FinalForest_Model2_Pre[!(FinalForest_Model2_Pre$parameter == "lprior"), ]
FinalForest_Model2_Pre <- FinalForest_Model2_Pre %>% 
  mutate(Model = "Model 2")

FinalForest_Model3_Pre <- FinalForest_Model3_Pre[!(FinalForest_Model3_Pre$parameter == "lprior"), ]
FinalForest_Model3_Pre <- FinalForest_Model3_Pre %>% 
  mutate(Model = "Model 3")

FinalForest_Model4_Pre <- FinalForest_Model4_Pre[!(FinalForest_Model4_Pre$parameter == "lprior"), ]
FinalForest_Model4_Pre <- FinalForest_Model4_Pre %>% 
  mutate(Model = "Model 4")


# Combine datasets
FinalForest_ModelsCombined_Pre <- rbind(FinalForest_Model1_Pre,FinalForest_Model2_Pre,FinalForest_Model3_Pre, FinalForest_Model4_Pre)
FinalForest_ModelsCombined_Pre <- FinalForest_ModelsCombined_Pre[!(FinalForest_ModelsCombined_Pre$parameter == "exp(lprior"), ]
FinalForest_ModelsCombined_Pre$Model <- as.factor(FinalForest_ModelsCombined_Pre$Model)
FinalForest_ModelsCombined_Pre$parameter <- gsub("CohortGeneralPopulation", "General Population", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("donation_count_2", "Number of blood donations (quadratic)", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("donation_count", "Number of blood donations", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("log_last_donation", "Days since last blood donation (log transformed)", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("blood_vol", "Blood volume (Nadler)", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("Smokingyes", "Smoking", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("MenstruationYes", "Menstruation", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("PreviousChildbirthyes", "Childbirth in the past", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- gsub("HormonalContraceptiones", "Hormonal contraception", FinalForest_ModelsCombined_Pre$parameter)
FinalForest_ModelsCombined_Pre$parameter <- as.factor(FinalForest_ModelsCombined_Pre$parameter)



desired_order <- c("General Population", "Age", "Blood volume (Nadler)", "Number of blood donations", "Number of blood donations (quadratic)", "Days since last blood donation (log transformed)", "Smoking", "Menstruation", "Childbirth", "Oral contraception")

plotdata_combined_M4_Pre <- FinalForest_ModelsCombined_Pre %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )

orlower_men <- -1
orupper_men <- 3

p <-  ggplot(plotdata_combined_M4_Pre, aes(x = m, y = parameter, color = parameter)) +
  geom_point(size=4) +
  geom_linerange(aes(xmin = ll, xmax = hh),
                 size = 1,
                 linetype = 1
                 ) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower_men,orupper_men)  +
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme_classic() +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        panel.grid.major.y = element_line(color = "gray", linetype = "dotted"),
        panel.grid.major.x = element_blank()
  ) +
  guides(shape = "none") +
  labs(title = "Logistic Models [Women - Premenopausal scaled]") +
  xlab("Coefficient") +
  ylab(" ") +
  facet_grid(Model ~ .,scale="free_y")

p
# 
FileName <- paste(FolderName, "/Forest Plot - Logistic - Final - Premenopausal scaled.png",sep="")
ggsave(FileName, width = 15, height = 10)


```


# Final Forest: POSTMENOPAUSAL WOMEN

```{r}
# Create final forest plot dataset
FinalForest_Model1_Post <- intervals_post
FinalForest_Model2_Post <- intervals_post_corrected
FinalForest_Model3_Post <- intervals_post_corrected_donhist
FinalForest_Model4_Post <- intervals_post_M4

FinalForest_Model1_Post <- FinalForest_Model1_Post[!(FinalForest_Model1_Post$parameter == "lprior"), ]
FinalForest_Model1_Post <- FinalForest_Model1_Post %>% 
  mutate(Model = "Model 1")

FinalForest_Model2_Post <- FinalForest_Model2_Post[!(FinalForest_Model2_Post$parameter == "lprior"), ]
FinalForest_Model2_Post <- FinalForest_Model2_Post %>% 
  mutate(Model = "Model 2")

FinalForest_Model3_Post <- FinalForest_Model3_Post[!(FinalForest_Model3_Post$parameter == "lprior"), ]
FinalForest_Model3_Post <- FinalForest_Model3_Post %>% 
  mutate(Model = "Model 3")

FinalForest_Model4_Post <- FinalForest_Model4_Post[!(FinalForest_Model4_Post$parameter == "lprior"), ]
FinalForest_Model4_Post <- FinalForest_Model4_Post %>% 
  mutate(Model = "Model 4")


# Combine datasets
FinalForest_ModelsCombined_Post <- rbind(FinalForest_Model1_Post,FinalForest_Model2_Post,FinalForest_Model3_Post, FinalForest_Model4_Post)
FinalForest_ModelsCombined_Post <- FinalForest_ModelsCombined_Post[!(FinalForest_ModelsCombined_Post$parameter == "exp(lprior"), ]
FinalForest_ModelsCombined_Post$Model <- as.factor(FinalForest_ModelsCombined_Post$Model)
FinalForest_ModelsCombined_Post$parameter <- gsub("CohortGeneralPopulation", "General Population", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("donation_count_2", "Number of blood donations (quadratic)", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("donation_count", "Number of blood donations", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("log_last_donation", "Days since last blood donation (log transformed)", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("blood_vol", "Blood volume (Nadler)", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("Smokingyes", "Smoking", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("MenstruationYes", "Menstruation", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("PreviousChildbirthyes", "Childbirth in the past", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- gsub("HormonalContraceptionyes", "Hormonal contraception", FinalForest_ModelsCombined_Post$parameter)
FinalForest_ModelsCombined_Post$parameter <- as.factor(FinalForest_ModelsCombined_Post$parameter)

desired_order <- c("General Population", "Age", "Blood volume (Nadler)", "Number of blood donations", "Number of blood donations (quadratic)", "Days since last blood donation (log transformed)", "Smoking", "Menstruation", "Childbirth in the past", "Hormonal contraception")
FinalForest_ModelsCombined_Post$parameter <- factor(FinalForest_ModelsCombined_Post$parameter, levels = rev(desired_order))


plotdata_combined_M4_Post <- FinalForest_ModelsCombined_Post %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )


orlower <- -1
orupper <- 20

p <-  ggplot(plotdata_combined_M4_Post, aes(x = m, y = parameter, color = parameter)) +
  geom_point(size=4) +
  geom_linerange(aes(xmin = ll, xmax = hh),
                 size = 1,
                 linetype = 1
                 ) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower,orupper)  +
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme_classic() +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        panel.grid.major.y = element_line(color = "gray", linetype = "dotted"),
        panel.grid.major.x = element_blank()
  ) +
  guides(shape = "none") +
  labs(title = "Logistic Models [Women - Postmenopausal scaled]") +
  xlab("Coefficient") +
  ylab(" ") +
  facet_grid(Model ~ .,scale="free_y")

p

FileName <- paste(FolderName, "/Forest Plot - Logistic - Final - Postmenopausal scaled.png",sep="")
ggsave(FileName, width = 15, height = 10)


```


# Final forest all groups dataset
```{r, fig8, width = 15, height = 10}
FinalForest_ModelsCombined_2 <- FinalForest_ModelsCombined
FinalForest_ModelsCombined_2$Group <- "Men"
FinalForest_ModelsCombined_Pre_2 <- FinalForest_ModelsCombined_Pre
FinalForest_ModelsCombined_Pre_2$Group <- "Women Premenopausal"
FinalForest_ModelsCombined_Post_2 <- FinalForest_ModelsCombined_Post
FinalForest_ModelsCombined_Post_2$Group <- "Women Postmenopausal"

FinalForest_ModelsCombined_All_2 <- rbind(FinalForest_ModelsCombined_2,FinalForest_ModelsCombined_Pre_2,FinalForest_ModelsCombined_Post_2)
FinalForest_ModelsCombined_All_2$Group <- factor(FinalForest_ModelsCombined_All_2$Group, levels = c("Men", "Women Premenopausal", "Women Postmenopausal"))
FinalForest_ModelsCombined_All_2$parameter <- gsub("Blood_volume", "Blood volume", FinalForest_ModelsCombined_All_2$parameter)


FinalForest_ModelsCombined_All_2$parameter <- gsub("MenstruationYes", "Menstruation", FinalForest_ModelsCombined_All_2$parameter)
FinalForest_ModelsCombined_All_2$parameter <- gsub("PreviousChildbirthyes", "Childbirth in the past", FinalForest_ModelsCombined_All_2$parameter)
FinalForest_ModelsCombined_All_2$parameter <- gsub("HormonalContraceptionyes", "Hormonal contraception", FinalForest_ModelsCombined_All_2$parameter)





desired_order_parameter <- c("General Population", "Age", "Blood volume (Nadler)", "Number of blood donations", "Number of blood donations (quadratic)", "Days since last blood donation (log transformed)", "Smoking", "Menstruation", "Childbirth in the past", "Hormonal contraception")
FinalForest_ModelsCombined_All_2$parameter <- factor(FinalForest_ModelsCombined_All_2$parameter, levels = rev(desired_order_parameter))


plotdata_ModelsCombined_All_2 <- FinalForest_ModelsCombined_All_2 %>% 
  mutate(
    ll = case_when(
      ll < orlower ~ orlower,
      TRUE ~ ll
    ),
    hh = case_when(
      hh > orupper ~ orupper,
      TRUE ~ hh
    )
  )

orlower_men <- -1
orupper_men <- 20

p <- ggplot(plotdata_ModelsCombined_All_2, aes(x = m, y = parameter, color = parameter)) +
  geom_point(size=4) +
  geom_linerange(aes(xmin = ll, xmax = hh),
                 size = 1,
                 linetype = 1
                 ) +
  scale_shape_manual(values = c(1,16)) +
  xlim(orlower_men,orupper_men)  +
  geom_vline(xintercept = 1 , color = "grey", linetype = "dotted",
             size = 1) +
  theme_classic() +
  theme(axis.title = element_text(size = 14),
        strip.text =  element_text(size = 14),
        legend.text = element_text(size = 12),
        strip.background = element_rect(fill = "white"),
        strip.background.x = element_blank(),
        axis.text.x = element_text(size = 14),
        axis.text.y = element_text(size = 14),
        panel.grid.major.y = element_line(color = "gray", linetype = "dotted"),
        panel.grid.major.x = element_blank(),
        legend.position = "none"
  ) +
  guides(shape = "none") +
  labs(title = "Logistic Models scaled") +
  xlab("Coefficient") +
  ylab(" ") +
  facet_grid(Model ~ Group,scale="free_y")



p
# 
FileName <- paste(FolderName, "/Forest Plot - Logistic - Final - All Groups scaled.png",sep="")
ggsave(FileName, width = 15, height = 10)




```



# Save Logistic regression Summary Data Dataset

```{r}
FinalForest_FIN_log <- FinalForest_ModelsCombined_All_2 %>% 
  add_column(Country = "FIN")

save(FinalForest_FIN_log, file = "~/proj/iron_measurement_comparisons/results/genpop_combined/low_ferritin_model_bayes/FinalForest_FIN_log_scaled.rdata")

saveRDS(FinalForest_ModelsCombined_All_2, file.path("~/proj/iron_measurement_comparisons/results/genpop_combined/low_ferritin_model_bayes/", "3. Cohort Comparison - Summary Data - Linear Regression scaled.rds"))


```
